Fix default ctor for structs with lists in C++
Updated C++ generator predicates to avoid erroneously omitting
default struct ctor if one of the fields is a collection of
immutable elements.

Moved field traversal predicate helpers to their respective
predicate classes to avoid misuse.

Added/updated smoke tests.

Resolves: #797
